pyproject.toml
[tool.poetry]
name = "agententerpriseval"
version = "0.1.0"
description = "AI Agent for Enterprise Valuation using Google ADK and EODHD Market Data"
authors = ["Matthew Tondreau <mmtondreau@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-adk = "^1.18.0"
python-dotenv = "^1.2.1"
hydra-core = "^1.3.2"
fastmcp = "^2.13.1"
httpx = "^0.28.1"
websockets = "^15.0.1"
psycopg2-binary = "^2.9.0"
sqlalchemy = "^2.0.0"
asyncpg = "^0.29.0"
greenlet = "^3.0.0"
overrides = "^7.7.0"


[tool.poetry.group.dev.dependencies]
black = "^25.11.0"
flake8 = "^7.3.0"
mypy = "^1.18.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
